requirements.txt
# requirements.txt
Flask
SQLAlchemy
requests
beautifulsoup4
python-docx
